+++ src/restaurants/restaurants.service.ts
@@ -3,6 +3,7 @@ import { InjectModel } from '@nestjs/mongoose';
 import * as mongoose from 'mongoose';
 import { Query } from 'express-serve-static-core';
 import { Restaurant } from './schemas/restaurant.schema';
+import APIFeatures from '../utils/apiFeatures.utils';
 
 @Injectable()
 export class RestaurantsService {
@@ -12,7 +13,8 @@ export class RestaurantsService {
   ){}
 
   async findAll(query: Query): Promise<Restaurant[]> {
-    const keyword = query.keyword ? {
+    const keyword = query.keyword
+      ? {
           name: {
             $regex: query.keyword,
             $options: 'i'
@@ -30,11 +32,18 @@ export class RestaurantsService {
   }
 
   async create(restaurant: Restaurant): Promise<Restaurant> {
-    const newRestaurant = await this.restaurantModel.create(restaurant); 
+    const location = await APIFeatures.getAddressLocation(restaurant.address);
+    const data = Object.assign(restaurant, { location });
+    const newRestaurant = await this.restaurantModel.create(data); 
     return newRestaurant;
   }
 
   async findById(id: string): Promise<Restaurant> {
+
+    if (!mongoose.isValidObjectId(id)) {
+      throw new NotFoundException('Invalid ID format, please enter a valid ID');
+    }
+
     const restaurant = await this.restaurantModel.findById(id);
 
     if (!restaurant) {
@@ -45,6 +54,11 @@ export class RestaurantsService {
   }
 
   async updateById(id: string, restaurant: Restaurant): Promise<Restaurant> {
+
+    if (!mongoose.isValidObjectId(id)) {
+      throw new NotFoundException('Invalid ID format, please enter a valid ID');
+    }
+
     return await this.restaurantModel.findByIdAndUpdate(id, restaurant, { 
       new: true,
       runValidators: true
@@ -52,6 +66,34 @@ export class RestaurantsService {
   }
 
   async deleteById(id: string): Promise<Restaurant> {
+
+    if (!mongoose.isValidObjectId(id)) {
+      throw new NotFoundException('Invalid ID format, please enter a valid ID');
+    }
+
     return await this.restaurantModel.findByIdAndDelete(id);
+  }
+
+  async uploadImages(id, files): Promise<any> {
+    const images = await APIFeatures.uploadFiles(files);
+
+    const restaurant = await this.restaurantModel.findByIdAndUpdate(
+      id,
+      {
+        images: images as Object[],
+      },{
+        new: true,
+        runValidators: true,
+      });
+
+    return restaurant;
+  }
+
+  async deleteImages(images): Promise<any> {
+    if (images.length === 0) {
+      return true;
+    }
+    const res = await APIFeatures.deleteFiles(images);
+    return res;
   }
 }
